==> notebooks/2_annotate_audio_files.ipynb <==
import sys
sys.path.append('../')

from typing import List, Dict

from src.config.project_paths import  get_data_file_path, \
     get_voxceleb2_metadata_path, get_voxceleb2_path, create_wav_dir_if_not_exists, get_project_root_path, get_voxceleb2_wav_path
import pandas as pd

## VoxCeleb 2

voxceleb2_metadata = pd.read_csv(get_voxceleb2_metadata_path(), sep=',')
voxceleb2_metadata = voxceleb2_metadata.drop(columns=["VGGFace2 ID "], axis=1).rename(
    columns={"VoxCeleb2 ID ": "user_id", "Gender ": "gender", "Set ": "set"})
voxceleb2_metadata["user_id"] = voxceleb2_metadata["user_id"].str.strip()
voxceleb2_metadata["gender"] = voxceleb2_metadata["gender"].str.strip()
voxceleb2_metadata["set"] = voxceleb2_metadata["set"].str.strip()

voxceleb2_metadata.head()

voxceleb2_path = get_voxceleb2_wav_path()
voxceleb2_path

id_paths = voxceleb2_path.iterdir()

import random

id_paths_list = list(voxceleb2_path.iterdir())
random_ids = random.sample(id_paths_list, 40)


annotations: List[Dict[str, any]] = []

import os

wav_file_path = get_voxceleb2_wav_path()

for id_path in random_ids:
    if id_path.stem not in [id_path.stem for id_path in random_ids]:
        continue
    create_wav_dir_if_not_exists(id_path.name)
    sample_paths = list(id_path.iterdir())
    
    for sample_path in sample_paths[:5]:
        wav_dir = os.path.join(get_voxceleb2_wav_path(), f"{id_path.name}/{sample_path.name}")
        os.makedirs(wav_dir, exist_ok=True)
        wav_paths = sample_path.iterdir()
        for wav_path in wav_paths:
            annotations.append(
                {
                    "wav_path": f"/data/vox2_test/wav/{id_path.name}/{sample_path.name}/{wav_path.stem}.wav",
                    "user_id": id_path.stem
                }
            )
            

len(annotations)

voxceleb2_user_to_metadata = voxceleb2_metadata[["user_id", "gender"]].drop_duplicates().set_index("user_id").to_dict(orient="index")

for annotation in annotations:
    annotation["gender"] = voxceleb2_user_to_metadata[annotation["user_id"]]["gender"]

pd.DataFrame(annotations).to_json(get_data_file_path("annotations_original.json"), orient="records", default_handler=str)
