==> notebooks/3_create_samples_source_annotations.ipynb <==
import sys
sys.path.append("../")
from src.config import project_paths
import pandas as pd
import os
from pathlib import Path

annotation_path = project_paths.get_data_file_path("annotations_original.json")

annotations = pd.read_json(annotation_path, orient="records")
len(annotations)

annotations.head()

user_counts = annotations['user_id'].value_counts()  
least_frequent_users = user_counts.nsmallest(100).index.tolist()

len(least_frequent_users)

least_frequent_users_40 = user_counts.nsmallest(100).index.tolist()
filtered_annotations_40 = annotations.apply(lambda row: row['user_id'] in least_frequent_users_40 and os.path.exists(Path(f"../{row['wav_path']}")), axis=1)
limited_annotations_40 = annotations[filtered_annotations_40]
limited_annotations_40.to_json(project_paths.get_data_file_path("annotations_400.json"), orient="records")
limited_annotations_40

least_frequent_users_20 = user_counts.nsmallest(20).index.tolist()
filtered_annotations_20 = annotations.apply(lambda row: row['user_id'] in least_frequent_users_20 and os.path.exists(Path(f"../../{row['wav_path']}")), axis=1)
limited_annotations_20 = annotations[filtered_annotations_20]
limited_annotations_20.to_json(project_paths.get_data_file_path("annotations_200.json"), orient="records")
limited_annotations_20

least_frequent_users_10 = user_counts.nsmallest(10).index.tolist()
filtered_annotations_10 = annotations.apply(lambda row: row['user_id'] in least_frequent_users_10 and os.path.exists(Path(f"../../{row['wav_path']}")), axis=1)
limited_annotations_10 = annotations[filtered_annotations_10]
limited_annotations_10.to_json(project_paths.get_data_file_path("annotations_100.json"), orient="records")
limited_annotations_10
