==> notebooks/6_preprocessing.ipynb <==
import sys
sys.path.append('../')

from collections import defaultdict
from pathlib import Path
from typing import List, Dict

from src.config.project_paths import get_voxceleb1_metadata_path, get_data_file_path, get_voxceleb_path, \
    get_project_root_path, get_voxceleb2_metadata_path, get_voxceleb2_path, create_wav_dir_if_not_exists, get_project_root_path, get_voxceleb2_wav_path
from src.annotate_data.annotate_from_path import data_path_to_annotation_list
import pandas as pd
from src.augmentation import add_noise
from src.config import project_paths
from pathlib import Path
import pandas as pd
import numpy as np
from copy import deepcopy
from typing import Dict, List
from uuid import uuid4
import os
import numpy as np
from pydub import AudioSegment
import soundfile as sf
from src.preprocessing import gaussian_noise_reduce
from tqdm.notebook import tqdm

voxceleb2_path = get_voxceleb2_wav_path()

import json

file_path = "../data/generated/annotations_multiply_amplitude.json"

with open(file_path, "r") as file:
    data = json.load(file)

for file in tqdm(data, total=len(data)):
    augmented_file_path = f"../{file['wav_path']}"
    gaussian_noise_reduce(augmented_file_path)
